Remove Tailwind and Angie
Keep a "snapshot" of the generated tailwind css and include it in the
prototype routes only. As soon as these routes are re-worked that
snapshot can just be deleted.

All other routes are KERN-only, no Tailwind. We remove Tailwind for now,
expecting that we don't really need it. Start small. We might add it
back later when we realize that we need it.